% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_functions.R
\name{logdet}
\alias{logdet}
\title{Take the log of the determinant of a matrix.}
\usage{
logdet(in_matrix)
}
\arguments{
\item{in_matrix}{A square, positive definite matrix.}
}
\value{
The log of the determinant of \code{in_matrix}.
}
\description{
Take the log of the determinant of a matrix.
}
